# specs/001-compliance-service/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Compliance Service API
  description: |
    Manages work authorizations, visas, permits, and employment compliance tracking.

    ## Features
    - Work authorization CRUD operations
    - Automated expiration monitoring and alerts
    - Compliance status reporting
    - Integration with Employee, Upload, and Notification services

    ## Base Path
    All endpoints are prefixed with `/compliance/v1`

  version: 1.0.0
  contact:
    name: MALIEV Engineering
    email: engineering@maliev.com
servers:
  - url: https://api.maliev.com/compliance/v1
    description: Production
  - url: https://staging-api.maliev.com/compliance/v1
    description: Staging
  - url: http://localhost:8080/compliance/v1
    description: Local Development
...
security:
  - BearerAuth: []
paths:
...
  /alerts:
    get:
      tags:
        - Alerts
      summary: Get compliance alerts
      description: Retrieves compliance alerts with optional filters
      operationId: getAlerts
      security:
        - BearerAuth: ['compliance.manage']
      parameters:
        - name: resolved
          in: query
          schema:
            type: boolean
            default: false
        - name: severity
          in: query
          schema:
            $ref: '#/components/schemas/AlertSeverity'
        - name: employeeId
          in: query
          schema:
            type: string
            format: uuid
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComplianceAlertResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /alerts/{alertId}/resolve:
    put:
      tags:
        - Alerts
      summary: Resolve compliance alert
      description: Marks a compliance alert as resolved with notes
      operationId: resolveAlert
      security:
        - BearerAuth: ['compliance.manage']
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveAlertRequest'
      responses:
        '200':
          description: Alert resolved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceAlertResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/AlertNotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ComplianceAlertResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        employeeId:
          type: string
          format: uuid
        employeeName:
          type: string
        alertType:
          $ref: '#/components/schemas/AlertType'
        severity:
          $ref: '#/components/schemas/AlertSeverity'
        message:
          type: string
        isResolved:
          type: boolean
        resolvedDate:
          type: string
          format: date-time
          nullable: true
        resolvedByName:
          type: string
          nullable: true
        resolutionNotes:
          type: string
          nullable: true
        createdDate:
          type: string
          format: date-time
    ResolveAlertRequest:
      type: object
      required:
        - resolutionNotes
      properties:
        resolutionNotes:
          type: string
          maxLength: 2000
...
    AlertType:
      type: string
      enum:
        - ExpirationWarning
        - Expired
        - DocumentMissing
        - VerificationRequired
        - RenewalReminder
    AlertSeverity:
      type: string
      enum:
        - Low
        - Medium
        - High
        - Critical
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: "INVALID_DATE_RANGE"
        message:
          type: string
          example: "Issue date cannot be after expiration date"
        details:
          type: object
          additionalProperties: true
  responses:
...
    Forbidden:
      description: Not authorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "NOT_AUTHORIZED"
            message: "User lacks required permissions"
...
    AlertNotFound:
      description: Alert not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "ALERT_NOT_FOUND"
            message: "Alert with the specified ID does not exist"
...